fix: race condition behavior: i810: EnterVT Failed nvidia: ioctl(...VT_WAITACTIVE...) block until switched to "X :2"'s tty
Description:
X :2 from Xconfigure.sh would start and switch to tty2, but in the meantime
Start_X would start and switch to tty7. In the nvidia case, the X :2 will wait
on that ioctl, but since the Start_X.sh script switched to tty7 already, it
will block. In the i810 case, EnterVT fails on the tty7 instance probably
going against the tty2 one.