Merges from charonmedia.org @ 1598. to lmce merge branch (mce-installer).
NOTE! This changes the debian/rules to not use the build-mce-installer.sh script because it relies on installing packages which a build script should not do, and also relies on files outside the repo. But we still need some portions of the build-mce-installer.sh, this needs to be fixed properly before this branch is merged to trunk.