Fixes a 64-bit issue preventing MythTV_PlugIn from marking programs set to record correctly.
The MythRecording struct defined a union of two 32 bit integers to a 64 bit key on 32 bit bit platform, but a union of two 64 bit integers with a 64 bit key on 64-bit platforms. So the key was now just the first integer, the chanid on 64 bit platforms. i.e. if you set anything for record on a channel, all recordings on that channel were marked for record in the orbiter.