Some tweaks to the MythTV fill db scripts.
First, we try to avoid a daily fill while we're in busy in the wizard by setting a timer.
Second, if a daily fill begins and then MythTvInitialFillDB.sh is run again we kill the daily fill. (But we make sure it is run again soon after the next mythbackend restart.)