Don't turn the TV on when music stops playing.
Chat log re the removed code:

(11:16:09 PM) Uplink_: thing is I commented that one out just now because it turned the screen on when you stopped playing music (which left the screen off)
(11:16:20 PM) Uplink_: and I see no use for that block of code
(11:16:39 PM) Uplink_: if the TV is on when you start playing media, it doesn't go off (even for music I think)
(11:17:09 PM) Uplink_: and if it's off when you start playing music, you want it to stay that way when you stop the music
(11:18:07 PM) MerkWerk: are you sure? maybe music is being stopped so that the user can use some other function
(11:19:40 PM) TSCHAKeee2: Uplink_: yeah, actually i had been looking for that chunk of code to take it out
(11:19:46 PM) TSCHAKeee2: i don't see a reason for it either.
(11:19:56 PM) TSCHAKeee2: not to mention, Possy will be _VERY_ happy
(11:20:03 PM) TSCHAKeee2: very very very very  very insanely happy
(11:20:14 PM) TSCHAKeee2: as he does not use the on-screen orbiter at all
(11:20:43 PM) Uplink_: OK, so out it is
(11:22:29 PM) ***totallymaxed is singing the same happy song as possy...:-)